functions/Get AI Hint/requirements.txt
appwrite>=21.0.0
google-generativeai>=0.8.0
